+++ ovenproof_scoreboard_plugin/scripts/mods/ovenproof_scoreboard_plugin/ovenproof_scoreboard_plugin_localization.lua
@@ -147,6 +147,12 @@ local localization = {
     track_ammo_crate_waste = {
         en = "Track Ammo Crate waste",
     },
+    track_ammo_crate_in_percentage = {
+        en = "Include Ammo Crates in total percentage of Ammo picked up",
+    },
+    track_ammo_crate_in_percentage_only_havoc = {
+        en = "Only include this if playing Havoc",
+    },
     -- Reusable labels
     row_kills = {
         en = "Kills",